renamed testimonials.js to carousel.js
