+++ scripts/carousel.js
