+++ scripts/index.js
@@ -4,4 +4,4 @@ import './header.js'
 import './mural.js'
 import './instaFeed.js'
 import './muralGallery.js'
-import './testimonials.js'
+import './carousel.js'
